FEAT: updated templates and generator to allow TF creation for DD
The generator library has a number of areas that are tightly couple to
the Sumo implementation. To ease the path toward "does this work" - I
temporarily removed support for DD in this commit. This can easily be
readded in the future. Likely a bit of refactoring needed to support
both.