FIX: corrected terraform template to reveal query from openslo def
The actual query from sample openslo DD definition was not making it to
the TF file. I removed the placeholder value from the TF template and
replaced with the appropriate variable. This worked as expected. I was
able to use the rendered TF files to create the SLO in DD.